test(mod_arith): widen interpreter test storage to i16
fixup for the verifier commit: !mod_arith.int<251 : i8> violates the new
storage-one-bit-wider-than-modulus invariant (251 needs 8 bits, so i8
storage is rejected). Use i16 storage instead.

+++ Test/Interpreter/ModArith/add.mlir
@@ -1,12 +1,12 @@
 // RUN: veir-interpret %s | filecheck %s
 
 "builtin.module"() ({
-  "func.func"() <{sym_name = "main", function_type = () -> !mod_arith.int<251 : i8>}> ({
-    %lhs = "mod_arith.constant"() <{ "value" = 250 : i8 }> : () -> !mod_arith.int<251 : i8>
-    %rhs = "mod_arith.constant"() <{ "value" = 10 : i8 }> : () -> !mod_arith.int<251 : i8>
-    %sum = "mod_arith.add"(%lhs, %rhs) : (!mod_arith.int<251 : i8>, !mod_arith.int<251 : i8>) -> !mod_arith.int<251 : i8>
-    "func.return"(%sum) : (!mod_arith.int<251 : i8>) -> ()
+  "func.func"() <{sym_name = "main", function_type = () -> !mod_arith.int<251 : i16>}> ({
+    %lhs = "mod_arith.constant"() <{ "value" = 250 : i16 }> : () -> !mod_arith.int<251 : i16>
+    %rhs = "mod_arith.constant"() <{ "value" = 10 : i16 }> : () -> !mod_arith.int<251 : i16>
+    %sum = "mod_arith.add"(%lhs, %rhs) : (!mod_arith.int<251 : i16>, !mod_arith.int<251 : i16>) -> !mod_arith.int<251 : i16>
+    "func.return"(%sum) : (!mod_arith.int<251 : i16>) -> ()
   }) : () -> ()
 }) : () -> ()
 
-// CHECK: Program output: #[0x09#8]
+// CHECK: Program output: #[0x0009#16]
